Initial Commit on move to Zola
This is a big dumb commit! There's a lot of things going on.

This is the barebones aphoenix.ca site. This includes a migration of all
content from WordPress except for the static images, which still need to
be grabbed for about 50% of the blog posts to make sense.

It includes some basic taxonomies - tags for photos, blogs, gear.

It includes some basic CSS, with special consideration for using custom
CSS Properties so that colour schemes are a snap later on. I've also
kind of sussed out the general colour scheme, and made it slightly
responsive.

There's some basic pagination for blog posts.