Introduce a shortcode for images
Image shortcode does the image preprocessing that needs to be done to serve images of an appropriate size. I have
elected for the following sizes:

- 320px wide and 180px tall for thumbnails
- 736px for full width (and auto height)
- 480px for float widths (and auto height)

This also updates the width for photobits.

I've added some styling for images to make them a bit prettier, and have updated only two entries - the Belwood
Lake photoblog entry and the hot-sauce-two blog entry. As part of this, ignore the processed files from the repo; I
don't think it's worthwhile to include them.

This progresses #14.